pyproject.toml
[project]
name = "google-genai-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "google-genai>=1.25.0",
    "python-dotenv>=1.1.1",
]
